set a type for the constant argument for min()

+++ drivers/block/blksnap/main.c
@@ -396,9 +396,11 @@ static int __init parameters_init(void)
 			 diff_storage_minimum);
 	}
 #ifdef CONFIG_64BIT
-	chunk_maximum_count_shift = min(40, chunk_maximum_count_shift);
+	chunk_maximum_count_shift = min((unsigned int)40,
+					chunk_maximum_count_shift);
 #else
-	chunk_maximum_count_shift = min(32, chunk_maximum_count_shift);
+	chunk_maximum_count_shift = min((unsigned int)32,
+					chunk_maximum_count_shift);
 #endif
 	return 0;
 }
